users_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
aio-pika
httpx
pytest
pytest-asyncio
pytest-cov